switch to using prettier-plugin-svelte for formatting
update to latest prettier plugin svelte
