+++ package.json
@@ -49,6 +49,7 @@
         "indent-string": "^3.2.0",
         "lodash": "^4.17.10",
         "prettier": "1.17.0",
+        "prettier-plugin-svelte": "0.4.0",
         "sinon": "^4.5.0",
         "source-map": "^0.7.3",
         "svelte": "3.2.0",
