+++ src/server.ts
@@ -13,7 +13,6 @@ import { HTMLPlugin } from './plugins/HTMLPlugin';
 import { CSSPlugin } from './plugins/CSSPlugin';
 import { wrapFragmentPlugin } from './api/wrapFragmentPlugin';
 import { TypeScriptPlugin } from './plugins/TypeScriptPlugin';
-import { set, get } from 'lodash';
 
 namespace TagCloseRequest {
     export const type: RequestType<
@@ -58,7 +57,6 @@ export function startServer() {
     });
 
     connection.onDidChangeConfiguration(({ settings }) => {
-        set(settings.svelte, 'plugin.html.format.settings', get(settings, 'html.format', {}));
         manager.updateConfig(settings.svelte);
     });
